Move DrawShared and DrawableShared to sub-module
